requirement.txt
openai>=1.12.0
pinecone>=3.0.0
langchain-community>=0.2.0
langchain-core>=0.1.34
pandas>=2.2.2
python-dotenv>=1.0.1
tqdm>=4.66.2
